day_18_Turtles/pyproject.toml
[project]
name = "day-18-turtles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorgram-py>=1.2.0",
]
